feat: minimize docker container
