+++ .dockerignore
@@ -1,11 +1,32 @@
-# Исключаем папку с CDK
+# CDK
 deploy/
 
-# Исключаем другие временные файлы и каталоги
-node_modules/
+# Node.js dependencies
+node_modules
 npm-debug.log
+yarn-error.log
+
+# Build output
+dist
+build
+coverage
+
+# Environment files
+.env
+
+# Dockerfile
 Dockerfile
 .dockerignore
+
+# Git files
 .git
 .gitignore
-.elasticbeanstalk/*
+
+# IDE and editor files
+.vscode
+.idea
+*.swp
+
+# OS-specific files
+.DS_Store
+Thumbs.db
