+++ package.json
@@ -7,7 +7,7 @@
   "license": "UNLICENSED",
   "scripts": {
     "prebuild": "rimraf dist",
-    "build": "nest build --webpack --webpackPath webpack.config.js",
+    "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
     "start:dev": "env-cmd -f .env nest start --watch",
@@ -22,28 +22,19 @@
   },
   "dependencies": {
     "@codegenie/serverless-express": "^4.14.1",
-    "@grpc/grpc-js": "^1.11.1",
-    "@grpc/proto-loader": "^0.7.13",
     "@nestjs/common": "^10.0.3",
-    "@nestjs/config": "^3.0.0",
     "@nestjs/core": "^10.0.3",
     "@nestjs/jwt": "^10.1.0",
     "@nestjs/microservices": "^10.3.10",
     "@nestjs/passport": "^10.0.0",
     "@nestjs/platform-express": "^10.0.3",
     "@nestjs/typeorm": "^10.0.2",
-    "amqp-connection-manager": "^4.1.14",
-    "amqplib": "^0.10.4",
-    "aws-cdk-lib": "^2.149.0",
     "aws-lambda": "^1.0.7",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
     "env-cmd": "^10.1.0",
     "helmet": "^7.0.0",
     "ioredis": "^5.4.1",
-    "kafkajs": "^2.2.4",
-    "mqtt": "^5.8.1",
-    "nats": "^2.28.0",
     "passport": "^0.6.0",
     "passport-http": "^0.3.0",
     "passport-jwt": "^4.0.1",
@@ -53,14 +44,12 @@
     "rxjs": "^7.8.1",
     "typeorm": "^0.3.20",
     "typeorm-naming-strategies": "^4.1.0",
-    "uuid": "^9.0.0"
+    "rimraf": "^5.0.1"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.0.3",
-    "@nestjs/schematics": "^10.0.1",
     "@nestjs/testing": "^10.0.3",
     "@types/express": "^4.17.17",
-    "@types/jest": "29.5.2",
     "@types/node": "^20.3.1",
     "@types/passport-jwt": "^3.0.8",
     "@types/passport-local": "^1.0.35",
@@ -69,20 +58,13 @@
     "@typescript-eslint/parser": "5.60.0",
     "eslint": "8.43.0",
     "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
     "jest": "29.5.0",
     "prettier": "^2.8.8",
-    "rimraf": "^5.0.1",
     "supertest": "^6.3.3",
-    "terser-webpack-plugin": "^5.3.10",
     "ts-jest": "29.1.0",
-    "ts-loader": "^9.5.1",
     "ts-node": "10.9.1",
-    "tsconfig-paths": "^4.2.0",
     "typescript": "^5.5.3",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-node-externals": "^3.0.0"
+    "webpack": "^5.93.0"
   },
   "jest": {
     "moduleFileExtensions": [
